only update camera and renderer aspects when it's necessary

+++ packages/framework/src/MouseWheelObserver.ts
@@ -4,7 +4,7 @@ import { passiveEventListener } from "./passiveEventListener";
 import type { EventBus } from "./EventBus.interface";
 import type { MouseWheelObserver as IMouseWheelObserver } from "./MouseWheelObserver.interface";
 
-export function MouseWheelObserver(htmlElement: HTMLElement, eventBus: EventBus, dimensionsState: Uint16Array, inputState: Int16Array): IMouseWheelObserver {
+export function MouseWheelObserver(htmlElement: HTMLElement, eventBus: EventBus, dimensionsState: Uint32Array, inputState: Int32Array): IMouseWheelObserver {
   let _zoomAmount = 0;
 
   function start(): void {
