+++ packages/framework/src/getPrimaryPointerVectorY.ts
@@ -2,7 +2,7 @@ import { Input } from "./Input";
 import { isPrimaryMouseButtonPressed } from "./isPrimaryMouseButtonPressed";
 import { isPrimaryTouchPressed } from "./isPrimaryTouchPressed";
 
-export function getPrimaryPointerVectorY(inputState: Int16Array): number {
+export function getPrimaryPointerVectorY(inputState: Int32Array): number {
   if (isPrimaryTouchPressed(inputState)) {
     return inputState[Input.code.T0_VECTOR_Y] / Input.vector_scale;
   }
