+++ packages/framework/src/isMousePointerInDimensionsBounds.ts
@@ -1,5 +1,5 @@
 import { Input } from "./Input";
 
-export function isMousePointerInDimensionsBounds(dimensionsState: Uint16Array, inputState: Int16Array): boolean {
+export function isMousePointerInDimensionsBounds(dimensionsState: Uint32Array, inputState: Int32Array): boolean {
   return inputState[Input.code.M_IN_BOUNDS] > 0;
 }
