+++ packages/framework/src/isPrimaryMouseButtonPressed.ts
@@ -1,5 +1,5 @@
 import { Input } from "./Input";
 
-export function isPrimaryMouseButtonPressed(inputState: Int16Array): boolean {
+export function isPrimaryMouseButtonPressed(inputState: Int32Array): boolean {
   return inputState[Input.code.M_BUTTON_L] > 0 && inputState[Input.code.M_IN_BOUNDS] > 0;
 }
