+++ packages/framework/src/isPrimaryPointerInDimensionsBounds.ts
@@ -1,6 +1,6 @@
 import { isMousePointerInDimensionsBounds } from "./isMousePointerInDimensionsBounds";
 import { isPromaryTouchInDimensionsBounds } from "./isPromaryTouchInDimensionsBounds";
 
-export function isPrimaryPointerInDimensionsBounds(dimensionsState: Uint16Array, inputState: Int16Array): boolean {
+export function isPrimaryPointerInDimensionsBounds(dimensionsState: Uint32Array, inputState: Int32Array): boolean {
   return isPromaryTouchInDimensionsBounds(dimensionsState, inputState) || isMousePointerInDimensionsBounds(dimensionsState, inputState);
 }
