+++ packages/framework/src/isPrimaryPointerPressed.ts
@@ -1,6 +1,6 @@
 import { isPrimaryMouseButtonPressed } from "./isPrimaryMouseButtonPressed";
 import { isPrimaryTouchPressed } from "./isPrimaryTouchPressed";
 
-export function isPrimaryPointerPressed(inputState: Int16Array): boolean {
+export function isPrimaryPointerPressed(inputState: Int32Array): boolean {
   return isPrimaryTouchPressed(inputState) || isPrimaryMouseButtonPressed(inputState);
 }
