+++ packages/framework/src/isPrimaryTouchPressed.ts
@@ -1,5 +1,5 @@
 import { Input } from "./Input";
 
-export function isPrimaryTouchPressed(inputState: Int16Array): boolean {
+export function isPrimaryTouchPressed(inputState: Int32Array): boolean {
   return inputState[Input.code.T0_PRESSURE] > 0 && inputState[Input.code.T0_IN_BOUNDS] > 0;
 }
