+++ packages/framework/src/isPromaryTouchInDimensionsBounds.ts
@@ -1,5 +1,5 @@
 import { Input } from "./Input";
 
-export function isPromaryTouchInDimensionsBounds(dimensionsState: Uint16Array, inputState: Int16Array): boolean {
+export function isPromaryTouchInDimensionsBounds(dimensionsState: Uint32Array, inputState: Int32Array): boolean {
   return inputState[Input.code.T0_IN_BOUNDS] > 0;
 }
