+++ packages/framework/src/updateCameraAspect.ts
@@ -2,7 +2,7 @@ import { updatePerspectiveCameraAspect } from "./updatePerspectiveCameraAspect";
 
 import type { Camera, PerspectiveCamera } from "three";
 
-export function updateCameraAspect(camera: Camera, dimensionsState: Uint16Array): void {
+export function updateCameraAspect(camera: Camera, dimensionsState: Uint32Array): void {
   if ("PerspectiveCamera" === camera.type) {
     return void updatePerspectiveCameraAspect(camera as PerspectiveCamera, dimensionsState);
   }
