+++ packages/framework/src/updatePerspectiveCameraAspect.ts
@@ -2,7 +2,7 @@ import { Dimensions } from "./Dimensions";
 
 import type { PerspectiveCamera } from "three";
 
-export function updatePerspectiveCameraAspect(camera: PerspectiveCamera, dimensionsState: Uint16Array): void {
+export function updatePerspectiveCameraAspect(camera: PerspectiveCamera, dimensionsState: Uint32Array): void {
   const aspect = dimensionsState[Dimensions.code.D_WIDTH] / dimensionsState[Dimensions.code.D_HEIGHT];
 
   if (camera.aspect === aspect) {
